Add and use new component for Payment block
